# Improve performance of `Oj.dump` with compat/rails mode