@Watson1978 2021-07-26
I retrieved a benchmark result on Intel Mac.

–                | before   | after    | result
--               | --       | --       | --
Oj.dump          | 1.435M   | 1.453M   | –
Oj.dump (compat) | 558.205k | 897.370k | 1.61x
Oj.dump (rails)  | 428.819k | 655.485k | 1.53x

## Env
- MacBook Pro (16-inch, 2019)
- macOS BigSur 11.5
- CPU 2.4 GHz 8cores Intel Core i9
- Ruby 3.0.2

## Before

```
Warming up --------------------------------------
             Oj.dump   139.896k i/100ms
    Oj.dump (compat)    55.788k i/100ms
     Oj.dump (rails)    43.009k i/100ms
Calculating -------------------------------------
             Oj.dump      1.435M (± 2.3%) i/s -      7.275M in   5.072783s
    Oj.dump (compat)    558.205k (± 2.5%) i/s -      2.789M in   5.000364s
     Oj.dump (rails)    428.819k (± 4.7%) i/s -      2.150M in   5.026970s
```

## After
```
Warming up --------------------------------------
             Oj.dump   142.370k i/100ms
    Oj.dump (compat)    89.373k i/100ms
     Oj.dump (rails)    66.401k i/100ms
Calculating -------------------------------------
             Oj.dump      1.453M (± 3.4%) i/s -      7.261M in   5.001952s
    Oj.dump (compat)    897.370k (± 1.9%) i/s -      4.558M in   5.081196s
     Oj.dump (rails)    655.485k (± 2.3%) i/s -      3.320M in   5.067834s
```