@ohler55 2021-08-02
I would not expect a significant change on non-trivial objects. For small it does help.